Firefox and Chrome JS report exception text as uppercase, while safari does not, so I had to modify the unit tests to convert everything to uppercase before the comparison.
Incidentally, there are two major ways to do case-insensitive string comparison. The first is the conversion of everything to upper or lower case then compare. That's great as long as there's no unicode. The "better" way is to use regex to do the compare (javascript's match function), but then you have to worry about escaping regex special characters.